+++ client/src/components/pages/Home/HomePresenter.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+//Template
+import HomeTemplate from '../../templates/HomeTemplate';
+
+function HomePresenter() {
+  console.log('got in homePresenter');
+  return <HomeTemplate />;
+}
+
+export default HomePresenter;
